Fixed some issues
Mostly fixed the issues with the Proficient Planning stuff for Genestealers and the Abominant. Tried to implement the Acolyte Hybrids, but that needs some more thought on my part.
I'm also giving up on the templates rework; it'll be more work than it's worth, so I'll stick with what works.