Implemented an Error Checker
Put in the ground work for the error checker. It currently lists the number of errors in the roster and prevents the user from saving rosters with errors. I want to do a tooltip to highlight the errors, but can't figure that out. This is most likely a temporary solution, but it fulfils my purposes for now.