Started Space Marine Relics
Did the first half of Space Marine Relics. Need to make sure that if a relic replaces a weapon that it restricts the datasheet to use the weapon it replaces. Space Marines are rough, since the large amount of characters means a lot of searching around for various weapon options.