Several Changes
Roles are now found in every datasheet to prepare for detachment restrictions. The basics of detachments is done, with allowing a roster to contain multiple detachments from different factions. Some testing will be needed, but that's for later/beta-testing. Also started with implementing Space Marine Chapter variants, in this case Ultramarines and the infrastructure for various Psychic Disciplines that the user may choose.