Big Dump; Finished Tyranids and AdMech
The past week has been terrible, so this is long overdue. Finished the remaining Tyranids datasheets and did all of AdMech, because I didn't want to do anything hard. The same will probably continue, so stratagems and layout overhaul are postponed until April. T'au is next, and it's going to be *fun*.