Gutting and Experimentation
Gutted out all of the factions except Space Marines, which will be my test faction since their subfactions are very simple. Discovered that I had had Template() under the Death Guard namespace, so that means a lot of clean up may have to be done. I'm trying to place all the code related to selecting a subfaction into the Faction class, but it's really stumping me right now so I'll sleep on it and see if I can get something working