Space Marines Working
At this moment, Space Marines are working. This was far more painful then it should've been, since I discovered my code for the different detachments in a roster was lackluster and missing apparently critical features. A start's been made though, and with Space Marines at least working means that most factions can use this as a starting point