Progress!
Was trying something I did a bit of research on and got it working! Datasheets and Faction now contains a method that adds an EventHandler to a ComboBox/ListBox respectively to dynamically change the background on specific items in the dropdown/list! Now I can restrict things without just hiding information!