Messing Around
Got some code working for the Malignant Plaguecaster, but I'm not sure if what I've done is a good idea, so the Stratagems branch has been created so I can mess with this. I'm happy with what I've done before hand with restrictions but with this a bit less so.