Refined Warlord and Relic Selections
Last time I was trying something that I wasn't happy with and today something just clicked and I refined it to something I do like. Started it with a few datasheets and some general bug fixing and code readability things. TODO was also updated with the list of datasheets affected by the Stratagems for future reference.