Beginning of Knights and Catching my Bearings
Been a long time since I last worked on this, mostly due to lack of interest. Getting back into it, it took me a while to remember where I left off. For reference sake, I was adding some sections of code (currently found in TODO) into Death Guard for Warlord and Relic restrictions across the army. Need to finish doing that. This overwhelmed me a bit, so I finished off the setup for Knights and did the first Armiger, which is incomplete but I'll get there. Freeblades will very much be an issue to tackle in the future once I divine a solution to that problem. I'll finish this some day!