Imperial Knights Nearly Complete
Implemented just about all of Imperial Knights. This codex is relatively easy, since there's so few entries and most of them are effectively duplicates of the same entry. All that remains is to finish relic selection and figure out Freeblades, which will probably be a new template, along with their Army of Renown.