Imperial Knights Complete
An easy book nearly complete! All that remains is Freeblade logic, which I'm going to hold out on for now since it's complicated. Next step is either Necrons update, finishing Death Guard or starting Leagues of Votann. I'm debating whether or not the restrictions that are currently in Death Guard are within the scope of this program.