Misc Changes and Leagues of Votann
Found a bug in Death Guard that should be fixed, it took far too long to. Continuing to have doubts on how far I want to take this program. I still intend to finish all the factions to the standard I currently have and at least add index restrictions to the weapon options on the datasheets (it looks far better than removing the option) but I haven't decided how much farther I'm going to go. Maybe this is just a funk, maybe a new course. We'll see.