Leagues of Votann
Submitting Leagues of Votann and some other miscellaneous changes. Hearthkyn Warriors took a long time to do, partly because I read the datasheet wrong the first time and tried doing a lot of it while under a sinus infection, so that didn't help. Next up is Chaos Space Marines, which is going to be a long one. Might finish Stratagems for Death Guard or implement restrictions in other datasheets, we'll see what I feel like. I am determined to finish this, regardless of how long it's going to take. This program is a point of pride for me, and I want to finish it cause I want this to be a spring board for future projects.