Beginning of CSM
Starting Chaos Space Marines and wow they have a lot of Warlord Traits and Relics. This is going to be a fun faction! Also renamed a bunch of Death Guard stuff since it overlaps with CSM and clarity is very important. Since CSM is technically the generic forces of Chaos, they get the name only while Death Guard gets the DG prefix.