Space Marines Revised
It's done! Been on a coding spree recently, finding the groove (and finally watching shows with the family again). Space Marines have been finished, with each datasheet double checked on points and slightly reworked where needed. Found a lot of bugs as well, or at least things that I may have outright missed. Stratagems are done for basic marines, now to move on to the Codex Supplements, which will mostly consist of character checks for relics and assuring that all the stratagems are named correctly. It's been a pain, but this was the hardest one!