Salamanders pt. 1
Started doing Salamanders relics and noticed that I forgot to do one of the Stratagems! Now each Character can have a chapter-unique Warlord Trait in addition to having a normal Warlord Trait. I've also double checked all Relics for the Salamanders and fixed a few bugs. On to Sergeant relics which should be easier!