Iron Hands and White Scars Done
Implemented and double checked relics for the Iron Hands and White Scars characters/sergeants along with the Iron Hands Dreadnought stratagem. Just have Imperial Fists left then I move on to Aeldari! I know I'll have to come back for some work for Deathwatch, DA, SW and BA, but it'll be nice to work on something else for a while.