Imperial Fists Done
And it's done! Imperial Fists have been implemented and Space Marines are *almost* done! I'll have to come back for the other supplements, but they're a bit more involved and interesting anyway. It'll be nice to work on something else for a change, though Aeldari could be a disaster as well. I'll find out soon!