Aeldari Codex Done
This update marks the completion of the Craftworlders and Harlequins! It's nice to not have to make many changes to get something working. I'll have to come for Ynnari, which has proven to be more complicated than expected. I don't like the solution, but Ynnari will be its own faction. The way it interacts with Drukhari means either a lot of modification to the Drukhari classes or just to copy and paste the hopefully soon-to-be updated Drukhari while making slight changes so they work with Ynnari. With that out of the way, next up is Tyranids!