T'au Complete
Title! Most of it was double-checking points, which were in error in a few places, but most of the effort came from Pathfinders and Strike/Breacher Teams with their various weapon options needing updating and Broadsides. Broadsides have weapon restrictions that I never implemented and it took me a bit to figure it out, but I did and everything works properly. Next is Orks, which hopefully isn't too bad. They'll be a pain with the whole weapon options rework, but not much stratagem implementation is required.