Grey Knights done
Grey Knights was not very difficult, but did result in a lot of simple fixes, including a few bugs that I implemented when coding. Next is Deathwatch, so back to Space Marines. I'm also starting to reach the point of code that I've only ever done once and haven't double checked, so more fixes may be inbound