Deatchwatch Prepared to Strike
This one took a while. The Kill Teams are the most complicated datasheets in the entire program and took a while to figure out, both in terms of what I had done and what needed to be fixed/applied. Everything should be correct, but there might be some hidden bugs in there. Next is Space Wolves, which shouldn't be too bad since most of their unique range is Characters and reskins of existing units. Shouldn't take nearly as long!