Space Wolves Redone
Space Wolves turned out to be significantly harder than I though, specifically Blood Claws, Grey Hunters, Skyclaws and Long Fangs, due to the ability for each of those units to be led by a Wolf Guard Battle Leader instead of the Sergeant but still keeping the Sergeant. The Code is messy and complicated (almost assuredly has bugs, but I only found one in testing that I couldn't replicate), but it is functional. Only a few Space Marine factions left to complete, but for now, something completely different. Next step is Drukhari, which will be more work since I remember having issues with some stuff and this is when I stopped implementing all the relics. Shouldn't be too bad, since nothing in Drukhari is radically different than what I've done before, but we'll see...